feat: add prune_notifications function to notification_store canister
- Introduced a new `prune_notifications` function to the Candid interface for managing notification cleanup.
- Removed the `set_pruning_timer` function from the initialization and post-upgrade processes to streamline notification management.
- Updated relevant module imports to reflect the removal of the pruning timer functionality.
