+++ src/canister/notification_store/src/api/mod.rs
@@ -1,3 +1,4 @@
 pub mod add_notification;
 pub mod mark_read;
 pub mod get_notifications;
+pub mod prune_notfications;
